requirements.txt
chromadb==0.4.22
docx2txt==0.8
json_repair==0.30.0
langchain==0.3.7
llama_index==0.11.22
ollama==0.3.3
numpy
pandas
plotly==5.18.0
pydantic==2.9.2
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
python_pptx==0.6.21
PyYAML==6.0.1
Requests==2.32.3
setuptools
streamlit==1.39.0
tqdm==4.66.1
umap_learn==0.5.5
unstructured[all]
pdfminer.six
pi-heif
unstructured-inference